requirements.txt
Django	4.0.4	4.0.4
asgiref	3.5.0	3.5.0
django-rest-framework	0.1.0	0.1.0
djangorestframework	3.13.1	3.13.1
pip	21.3.1	22.0.4
pytz	2022.1	2022.1
setuptools	60.2.0	62.1.0
sqlparse	0.4.2	0.4.2
tzdata	2022.1	2022.1
wheel	0.37.1	0.37.1
Django~=4.0.4
djangorestframework~=3.13.1